pyproject.toml
[project]
name = "agenticrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-huggingface>=0.0.1",
    "langchain-text-splitters>=0.0.1",
    "huggingface-hub>=0.19.0",
    "transformers>=4.36.0",
    "torch>=2.0.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "ragas>=0.4.3",
    "rank-bm25>=0.2.2",
    "scikit-learn>=1.3.0",
    "streamlit>=1.28.0",
    "google-generativeai>=0.3.0",
]
